requirements.txt
fastapi
uvicorn
requests
cryptography
pyjwt
colorama